fix: in io strings distance over sea is an int

+++ src/HCMUtils/Types/LegacyInput.cs
@@ -38,7 +38,7 @@ public record LegacyInput(
     bool ChannelOccupation,
     Temperature? SeaTemperature,
     int TxServiceAreaRadius,
-    double? DistanceOverSea,
+    int? DistanceOverSea,
     string TxEmissionDesignation,
     double? PermissibleFieldStrength,
     Country TxCountry,
